# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: DreamBoard API Contract
  version: 1.2.0
  description: |
    Unified API contract for DreamBoard web and backend.
    Endpoints marked with `x-implemented: false` are planned and documented to prevent client/backend drift.
servers:
  - url: http://localhost:3000
  - url: http://45.139.77.52:3000

tags:
  - name: auth
  - name: boards
  - name: templates
  - name: cards
  - name: versions
  - name: share-links
  - name: uploads
  - name: export

paths:
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh token
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'

  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'

  /v1/boards:
    get:
      tags: [boards]
      summary: List boards
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
    post:
      tags: [boards]
      summary: Create board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'

  /v1/boards/recent:
    get:
      tags: [boards]
      summary: List recently opened boards (UX Sprint Day 1)
      description: Returns recent boards ordered by lastOpenedAt DESC; only boards with non-null lastOpenedAt.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
        - in: query
          name: q
          description: Case-insensitive search against board title.
          schema:
            type: string
            maxLength: 120
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'

  /v1/boards/from-template:
    post:
      tags: [boards, templates]
      summary: Create board from template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardFromTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Template not found

  /v1/boards/{boardId}:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    get:
      tags: [boards]
      summary: Get board
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Not found
    patch:
      tags: [boards]
      summary: Update board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
    delete:
      tags: [boards]
      summary: Delete board
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteBoardResponse'

  /v1/boards/{boardId}/pin:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    post:
      tags: [boards]
      summary: Pin board on Home Dashboard
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Board not found
    delete:
      tags: [boards]
      summary: Unpin board on Home Dashboard
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: Board not found

  /v1/templates:
    get:
      tags: [templates]
      summary: List board templates for onboarding/template picker
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesResponse'

  /v1/boards/{boardId}/cards:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    get:
      tags: [cards]
      summary: List cards
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
    post:
      tags: [cards]
      summary: Create card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCardResponse'

  /v1/boards/{boardId}/cards/{cardId}:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/CardId'
    patch:
      tags: [cards]
      summary: Update card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
    delete:
      tags: [cards]
      summary: Delete card
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCardResponse'

  /v1/boards/{boardId}/cards/reorder:
    post:
      tags: [cards]
      summary: Reorder cards (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/versions:
    get:
      tags: [versions]
      summary: List versions (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build
    post:
      tags: [versions]
      summary: Create snapshot version (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/versions/{versionId}/restore:
    post:
      tags: [versions]
      summary: Restore version (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/share-links:
    get:
      tags: [share-links]
      summary: List share links (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build
    post:
      tags: [share-links]
      summary: Create share link (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/share-links/{linkId}:
    delete:
      tags: [share-links]
      summary: Revoke share link (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/uploads/intents:
    post:
      tags: [uploads]
      summary: Create upload intent (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/uploads/finalize:
    post:
      tags: [uploads]
      summary: Finalize upload (planned)
      x-implemented: false
      responses:
        '501':
          description: Not implemented in current backend build

  /v1/boards/{boardId}/export/png:
    get:
      tags: [export]
      summary: Export PNG
      description: Client-side export in current build (no backend API endpoint).
      x-implemented: false
      responses:
        '501':
          description: Client-side only in current build

  /v1/boards/{boardId}/export/jpg:
    get:
      tags: [export]
      summary: Export JPG
      description: Client-side export in current build (no backend API endpoint).
      x-implemented: false
      responses:
        '501':
          description: Client-side only in current build

components:
  parameters:
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: string

  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string

    LoginResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            email:
              type: string

    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string

    LogoutResponse:
      type: object
      properties:
        message:
          type: string

    Board:
      type: object
      required: [id, title, createdAt, updatedAt, isPinned]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        isPinned:
          type: boolean
        lastOpenedAt:
          type: string
          format: date-time
          nullable: true
        templateId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateBoardRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string

    UpdateBoardRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string

    DeleteBoardResponse:
      type: object
      properties:
        deleted:
          type: boolean
        board:
          $ref: '#/components/schemas/Board'

    Card:
      type: object
      required: [id, boardId, type, text, createdAt, updatedAt]
      properties:
        id:
          type: string
        boardId:
          type: string
        type:
          type: string
          enum: [text]
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateCardRequest:
      type: object
      properties:
        type:
          type: string
          enum: [text]
        text:
          type: string

    CreateCardResponse:
      type: object
      properties:
        created:
          $ref: '#/components/schemas/Card'

    UpdateCardRequest:
      type: object
      properties:
        text:
          type: string

    DeleteCardResponse:
      type: object
      properties:
        deleted:
          type: boolean
        card:
          $ref: '#/components/schemas/Card'

    BoardListItem:
      allOf:
        - $ref: '#/components/schemas/Board'
        - type: object
          properties:
            isPinned:
              type: boolean
              default: false
            lastOpenedAt:
              type: string
              format: date-time
              nullable: true
            pinUpdatedAt:
              type: string
              format: date-time
              nullable: true

    RecentBoardsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BoardListItem'
        nextCursor:
          type: string
          nullable: true

    PinBoardRequest:
      type: object
      required: [pinned]
      properties:
        pinned:
          type: boolean
          description: true to pin board, false to unpin.

    PinBoardResponse:
      type: object
      required: [boardId, pinned]
      properties:
        boardId:
          type: string
        pinned:
          type: boolean
        pinUpdatedAt:
          type: string
          format: date-time
          nullable: true

    Template:
      type: object
      required: [id, name, snapshot, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        snapshot:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TemplatesResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Template'

    CreateBoardFromTemplateRequest:
      type: object
      required: [templateId]
      properties:
        templateId:
          type: string
        title:
          type: string
          description: Optional custom title. Falls back to template/snapshot title.